Add styleguidist for isolated UI development
