+++ electron-react/src/StyleguidistMuiWrapper.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { ThemeProvider } from "@material-ui/core/styles";
+import RsgWrapper from "react-styleguidist/lib/client/rsg-components/Wrapper/Wrapper";
+
+import muiTheme from "./muiTheme";
+import "./assets/css/App.css";
+
+const RsgMuiWrapper = ({ children, ...rest }) => (
+  <RsgWrapper {...rest}>
+    <ThemeProvider theme={muiTheme}>{children}</ThemeProvider>
+  </RsgWrapper>
+);
+
+export default RsgMuiWrapper;
