+++ electron-react/styleguide.config.js
@@ -1,0 +1,13 @@
+const path = require("path");
+const { version } = require("./package");
+
+module.exports = {
+  components: "src/components/**/[A-Z]*.js",
+  styleguideComponents: {
+    Wrapper: path.join(__dirname, "src/StyleguidistMuiWrapper"),
+  },
+  ribbon: {
+    url: "https://github.com/Chia-Network/chia-blockchain",
+  },
+  version,
+};
